InputView의 분리
- 별도의 의미를 가지는 구조체여서 별도의 파일로 분리
- get이라는 어울리지 않는 함수명 변경
