+++ RacingCar/RacingCar.swift
@@ -7,28 +7,8 @@
 
 import Foundation
 
-struct InputView {
-    
-    static func getCarCount() -> Int {
-        print("자동차 대수는 몇 대인가요? ", terminator: "")
-        guard let input = readLine(), let carCount = Int(input) else {
-            print("입력값이 숫자가 아닙니다.")
-            return -1
-        }
-        return carCount
-    }
-    
-    static func getTryCount() -> Int {
-        print("시도할 횟수는 몇 회인가요? ", terminator: "")
-        guard let input = readLine(), let count = Int(input) else {
-            print("입력값이 숫자가 아닙니다.")
-            return -1
-        }
-        return count
-    }
-}
-
 struct ResultView {
+    
     static func showResult(cars: [Car]) {
         for i in 0..<cars.count {
             cars[i].printDistance()
@@ -44,8 +24,8 @@ struct RacingCar {
     
     /// UI
     mutating func prepareRace() {
-        genereteCar(input: InputView.getCarCount())
-        tryCount = InputView.getTryCount()
+        genereteCar(input: InputView.insertCarCount())
+        tryCount = InputView.insertTryCount()
     }
     
     func showResult() {
